+++ cpp/tests/replace/replace_tests.cpp
@@ -580,9 +580,7 @@ TEST_F(ReplaceDictionaryTest, StringsKeys)
 
   auto result =
     cudf::find_and_replace_all(input->view(), values_to_replace->view(), replacements->view());
-  cudf::test::print(result->view());
   auto decoded = cudf::dictionary::decode(result->view());
-  cudf::test::print(decoded->view());
   cudf::test::strings_column_wrapper expected({"z", "b", "z", "c", "b", "z", "c", "b"});
 
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(*decoded, expected);
@@ -591,24 +589,32 @@ TEST_F(ReplaceDictionaryTest, StringsKeys)
 TEST_F(ReplaceDictionaryTest, InputAndReplacementNulls)
 {
   cudf::test::fixed_width_column_wrapper<int32_t> input_w({1, 2, 1, 2, 0, 3, 4, 4, 3},
-                                                          {1, 1, 1, 1, 1, 1, 1, 1, 1});
+                                                          {1, 1, 1, 1, 0, 1, 1, 1, 1});
   auto input = cudf::dictionary::encode(input_w);
   cudf::test::fixed_width_column_wrapper<int32_t> values_to_replace_w({2, 3});
   auto values_to_replace = cudf::dictionary::encode(values_to_replace_w);
-  cudf::test::print(values_to_replace->view());
-  cudf::test::fixed_width_column_wrapper<int32_t> replacements_w({5, 0}, {1, 1});
+  cudf::test::fixed_width_column_wrapper<int32_t> replacements_w({5, 0}, {1, 0});
   auto replacements = cudf::dictionary::encode(replacements_w);
-  cudf::test::print(replacements->view());
 
   auto result =
     cudf::find_and_replace_all(input->view(), values_to_replace->view(), replacements->view());
-  cudf::test::print(result->view());
   auto decoded = cudf::dictionary::decode(result->view());
-  cudf::test::print(decoded->view());
   cudf::test::fixed_width_column_wrapper<int32_t> expected({1, 5, 1, 5, 0, 0, 4, 4, 0},
                                                            {1, 1, 1, 1, 0, 0, 1, 1, 0});
 
   CUDF_TEST_EXPECT_COLUMNS_EQUAL(*decoded, expected);
+}
+
+TEST_F(ReplaceDictionaryTest, EmptyReplacement)
+{
+  cudf::test::fixed_width_column_wrapper<double> input_w(
+    {1.0, 2.0, 1.0, 2.0, 0.0, 3.0, 4.0, 4.0, 3.0}, {1, 1, 1, 1, 0, 1, 1, 1, 1});
+  auto input = cudf::dictionary::encode(input_w);
+  cudf::test::fixed_width_column_wrapper<double> empty_w({});
+  auto empty  = cudf::dictionary::encode(empty_w);
+  auto result = cudf::find_and_replace_all(input->view(), empty->view(), empty->view());
+
+  CUDF_TEST_EXPECT_COLUMNS_EQUAL(*result, *input);
 }
 
 CUDF_TEST_PROGRAM_MAIN()